Marujo

+++ src/controllers/userController.js
@@ -6,15 +6,30 @@ const path = require('path');
 const userService = require('../service/userService');
 const validator = require('../utils/userValidator');
 
-const getPerfil = async (req, res) => {
-    try {
-        let user_id;
+const getPerfil = (req, res) => {
+    const { id, name, username, email, biography, image, created_at, updated_at } = req.user
 
-        if(req.params.user_id){
-            user_id = req.params.user_id;
-        }else {
+    return res.status(200).json(
+        {
+            id,
+            name,
+            username,
+            email,
+            biography,
+            image,
+            created_at,
+            updated_at
+        }
+    )
+}
+
+const getPerfilById = async (req, res) => {
+    try {
+        let user_id = req.params.user_id;
+
+        if (user_id === "0") {
             user_id = req.cookies.session_id;
-            user_id = jwt.verify(user_id, config.SECRET_KEY);
+            user_id = jwt.verify(user_id, SECRET_KEY);
             user_id = user_id.user.id;
         }
 
@@ -85,11 +100,23 @@ const registerUser = async (req, res) => {
 
 const updatePerfil = async (req, res) => {
     try {
-        const { name, username, biography } = req.body;
+        const { name, username, biography } = JSON.parse(req.body.data);
 
         const updates = {};
-        if (name) updates.name = name.trim();
-        if (username) updates.username = username.trim();
+        if (!validator.isEmpty(name)) { updates.name = name.trim() }
+        else {
+            const error = new Error("Nome não pode estar vazio.");
+            error.status = 400;
+            throw error;
+        };
+
+        if (!validator.isEmpty(username)) { updates.username = username.trim() }
+        else {
+            const error = new Error("Nome de usuário não pode estar vazio.");
+            error.status = 400;
+            throw error;
+        }
+
         if (biography) updates.biography = biography.trim();
 
         if (req.file) {
@@ -97,12 +124,6 @@ const updatePerfil = async (req, res) => {
         }
 
         if (updates.name) {
-            if (validator.isEmpty(updates.name)) {
-                const error = new Error("Nome não pode estar vazio.");
-                error.status = 400;
-                throw error;
-            }
-
             if (!validator.nameSize(updates.name)) {
                 const error = new Error("O nome não pode ser maior que 100 caracteres.");
                 error.status = 422;
@@ -111,12 +132,6 @@ const updatePerfil = async (req, res) => {
         }
 
         if (updates.username) {
-            if (validator.isEmpty(updates.username)) {
-                const error = new Error("Nome de usuário não pode estar vazio.");
-                error.status = 400;
-                throw error;
-            }
-
             if (validator.hasSpace(updates.username)) {
                 const error = new Error("Nome de usuário não pode conter espaços.");
                 error.status = 400;
@@ -149,7 +164,7 @@ const updatePerfil = async (req, res) => {
         const oldImage = await userService.updatePerfilService(user_id, updates);
 
         if (updates.image) {
-            if (oldImage !== "DEFAULT.png") {
+            if ((oldImage !== "DEFAULT.png") && oldImage) {
                 const image = path.join(__dirname, '..', 'uploads', oldImage);
                 fs.unlink(image, (err) => {
                     if (err) {
@@ -178,7 +193,7 @@ const updatePerfil = async (req, res) => {
     }
 }
 
-const checkUsername = async (req, res, next) => {
+const checkUsername = async (req, res) => {
     try {
         const username = req.params.username;
 
@@ -263,6 +278,7 @@ const login = async (req, res) => {
 
 module.exports = {
     getPerfil,
+    getPerfilById,
     checkUsername,
     checkEmail,
     registerUser,
